tests/hetzner: Finish expr for the coordinator.
This finally enables the TEST_MODE we mentioned in an earlier commit. In
addition, we need to make sure we have all dependencies that might need
networking access to the coordinator machine for the build processes to
finish flawlessly.

One thing I don't like here is the amount of memory we need for the
coordinator, but it already is so low that no OOM conditions arise
anymore, at least on my machine.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>