Keep track of and set system.stateVersion
NixOps now sets the default value of system.stateVersion to the NixOS
version in use when the machine was first created. For instance, if
the machine was created during the 14.12 era, then we add

  system.stateVersion = mkDefault "14.12";

to the generated physical spec. (Of course, you can always override
system.stateVersion in your configuration.)

This ensures in particular that SSH host keys of existing deployments
don't break during the 14.12 -> 15.09 transition.